+++ packages/next/src/shared/lib/lazy-dynamic/loadable.tsx
@@ -1,7 +1,7 @@
 import { Suspense, lazy } from 'react'
 import { BailoutToCSR } from './dynamic-bailout-to-csr'
 import type { ComponentModule } from './types'
-import { PreloadCss } from './preload-css'
+import { PreloadChunks } from './preload-chunks'
 
 // Normalize loader to return the module as form { default: Component } for `React.lazy`.
 // Also for backward compatible since next/dynamic allows to resolve a component directly with loader
@@ -52,7 +52,7 @@ function Loadable(options: LoadableOptions) {
       <>
         {/* During SSR, we need to preload the CSS from the dynamic component to avoid flash of unstyled content */}
         {typeof window === 'undefined' ? (
-          <PreloadCss moduleIds={opts.modules} />
+          <PreloadChunks moduleIds={opts.modules} />
         ) : null}
         <Lazy {...props} />
       </>
